broken monolith into /services and then each service based on function. library is in /lib. will need to publish lib package to run microservices
