+++ main_test.go
@@ -1,13 +1,14 @@
-package gojenga
+package main
 
 import (
 	"fmt"
+	"gojenga/gjServer"
 	"testing"
 )
 
 func TestOne(t *testing.T) {
 	fmt.Println("testOne running")
-	if testingFunc() == true {
+	if gjServer.testingFunc() == true {
 		t.Fatal("Throwing an error from testOne")
 	} else {
 		fmt.Println("test succeeded no errors")
