+++ services/gjDelete/main.go
@@ -1,0 +1,60 @@
+package gjDelete
+
+import (
+	"context"
+	"errors"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/exporters/jaeger"
+	"go.opentelemetry.io/otel/sdk/resource"
+	tracesdk "go.opentelemetry.io/otel/sdk/trace"
+	semconv "go.opentelemetry.io/otel/semconv/v1.7.0"
+	"gojenga/lib/gjLib"
+)
+
+const (
+	service     = "blockchain"
+	environment = "alpha"
+	id          = 1
+	verion      = "1.0.10"
+)
+
+func TracerProvider(url string) (*tracesdk.TracerProvider, error) {
+	// Create the Jaeger exporter
+	exp, err := jaeger.New(jaeger.WithCollectorEndpoint(jaeger.WithEndpoint(url)))
+	if err != nil {
+		return nil, err
+	}
+	tp := tracesdk.NewTracerProvider(
+		// Always be sure to batch in production.
+		tracesdk.WithBatcher(exp),
+		// Record information about this application in a Resource.
+		tracesdk.WithResource(resource.NewWithAttributes(
+			semconv.SchemaURL,
+			semconv.ServiceNameKey.String(service),
+			attribute.String("environment", environment),
+			attribute.Int64("ID", id),
+		)),
+	)
+	return tp, nil
+}
+
+func DeleteUser(jsonResponse gjLib.Traffic, ctx context.Context) (string, error) {
+	tr := otel.Tracer("crypto-trace")
+	ctx, span := tr.Start(ctx, "deleteUser")
+	span.SetAttributes(attribute.Key("testset").String("value"))
+	defer span.End()
+	r := gjLib.RunDynamoDeleteItem("ledger", jsonResponse.SourceAccount)
+	if r["code"] == "1" {
+		return "--> " + r["msg"], errors.New("--> " + r["msg"])
+	}
+	gjLib.RunDynamoDeleteItem("users", jsonResponse.SourceAccount)
+	if r["code"] == "1" {
+		return "--> " + r["msg"], errors.New("--> " + r["msg"])
+	}
+
+	//deleteMongo(jsonResponse, ctx)
+	//jsonResponse.Role = "USER"
+	//deleteMongo(jsonResponse, ctx)
+	return "Delete Succesful", errors.New("Delete Succesful")
+}
